refactor: 테마 커스텀 레포지토리 생성
