+++ src/main/java/roomescape/theme/infrastructure/JpaThemeRepository.java
@@ -1,0 +1,7 @@
+package roomescape.theme.infrastructure;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import roomescape.theme.domain.Theme;
+
+public interface JpaThemeRepository extends JpaRepository<Theme, Long>, ThemeCustomRepository {
+}
